refactor(app): move App component into containers dir
