+++ src/containers/App.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react'
 import {bindActionCreators} from 'redux'
 import {connect} from 'react-redux'
-import Counter from './components/Counter'
-import * as CounterActions from './actions/counter'
+import Counter from '../components/Counter'
+import * as CounterActions from '../actions/counter'
+import WidgetList from './WidgetList'
 
 function mapStateToProps(state) {
   return {
@@ -19,10 +20,7 @@ class App extends Component {
     return (
       <div>
         <h1>Hello, World</h1>
-        <Counter
-          counter={this.props.counter}
-          increment={this.props.increment}
-          decrement={this.props.decrement} />
+        <WidgetList />
       </div>
     )
   }
