+++ src/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux'
 
-import App from './App';
+import App from './containers/App';
 import configureStore from './store/configureStore'
 
 const store = configureStore()
